Bugfix casts to correct data types

+++ concept_A/MediaFusionGCV/MediaFusionGCV_API.h
@@ -37,14 +37,14 @@ struct deviceProperties
 
 
 	MEDIAFUSIONGCV_API errorState mediaLib_GStreamerInit(int argc, char* argv[]);
-	MEDIAFUSIONGCV_API int32_t mediaLib_create(SourceType, SinkType, const char*);
+	MEDIAFUSIONGCV_API size_t mediaLib_create(SourceType, SinkType, const char*);
 	MEDIAFUSIONGCV_API errorState mediaLib_init(int32_t, const char*, const char*);
-	MEDIAFUSIONGCV_API errorState mediaLib_getDevices(int32_t, int32_t&, deviceProperties**);
+	MEDIAFUSIONGCV_API errorState mediaLib_getDevices(int32_t, size_t&, deviceProperties**);
 	MEDIAFUSIONGCV_API errorState mediaLib_setDevice(int32_t, int32_t, int32_t);
 	MEDIAFUSIONGCV_API errorState mediaLib_startStreaming(int32_t );
 	MEDIAFUSIONGCV_API errorState mediaLib_stopStreaming(int32_t );
 
-	MEDIAFUSIONGCV_API int32_t mediaLib_delete(int32_t);
+	MEDIAFUSIONGCV_API size_t mediaLib_delete(int32_t);
 
 
 #ifdef __cplusplus
